refactor(web): extract schema/config rendering into web_schema (web_server.py god-file slice R1-C6)
